+++ Instructions/STY.c
@@ -7,7 +7,7 @@
 //doing these in the order they appear in the OpCode Matrix
 
 //zp
-void	STY_ZP(Registers *pRegs, MemModule *pMem, uint8_t argAddr)
+void	STY_ZP(Registers *pRegs, MemModule *pMem, uint16_t argAddr)
 {
 	WriteZP(pMem, argAddr, pRegs->Y);
 }
@@ -19,7 +19,7 @@ void	STY_Absolute(Registers *pRegs, MemModule *pMem, uint16_t argAddr)
 }
 
 //zp,x
-void	STY_ZPX(Registers *pRegs, MemModule *pMem, uint8_t argAddr)
+void	STY_ZPX(Registers *pRegs, MemModule *pMem, uint16_t argAddr)
 {
 	WriteZP(pMem, argAddr + pRegs->X, pRegs->Y);
 }
